Add one basic unit test
Since the Java skeleton starts out with a unit test to check that the
player returns a number I thought it would be fair if the JavaScript
player started with a test too.

I've added the mocha & chai development dependencies for testing locally
(won't be installed in production) and added a "test script" so that you
can conveniently type `npm test` to run the tests.

The tests checks that the Player's betRequest method always returns a
number, even when handed an empty game state.